% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lfcShrink.R
\name{add_lfcShrink}
\alias{add_lfcShrink}
\title{Add LFC Shrink}
\usage{
add_lfcShrink(myResults, dds, myContrast)
}
\arguments{
\item{myResults}{DESeq results object}

\item{dds}{DESeq Object}

\item{myContrast}{coefficient or contrast that you used to}
}
\description{
Takes in DESeq results object and turns it into a nice data frame, complete with column for shrunken lfcs.
}
\details{
Returns data frame with DESeq results entries for each gene (significant and non-), with separate shrunken lfc column.

Generated in a rage when I got frustrated with trying to pass multiple objects in a list out of an R function.
One day, this may get folded into deseq_coef or deseq_contrasts.
}
\examples{
results_sex_df <- add_lfcShrink(results_sex, dds_sva, c('Sex', 'M', 'F'))
}
\keyword{DESeq}
