% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSVs.R
\name{checkSVs}
\alias{checkSVs}
\title{Check Surrogate Variables}
\usage{
checkSVs(MyDESeqDataSetVar, mySVseq, nSurr)
}
\arguments{
\item{nSurr}{number of surrogate variables contained within SVA package that you want to plot}

\item{myDESeqDataSetVar}{Imported DESeq data, dollar sign, sampleTable name you want to plot}

\item{svseq}{svseq object created with SVA package}
}
\description{
Do your surrogate variables correlate to any known sources of variance in your dataset?
This function will help you visualize that.
}
\examples{
checkSVs(myData$Sex, svseq, 3)
}
\keyword{Batch}
\keyword{Correction}
\keyword{SVA}
