% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_coef.R
\name{deseq_coef}
\alias{deseq_coef}
\title{You could just call DESeq() in your code... or you could call this nifty function, which 
also saves the output to a table AND an R variable for later use.}
\usage{
deseq_coef(dds, myAlpha, coefficient, nickname, currDate, expName,
  amySexGenes)
}
\arguments{
\item{dds}{DESeq object}

\item{myAlpha}{number - threshold for calling genes significant}

\item{coefficient}{coefficient for which you want DE gene list. To see all available, enter matrix(resultsNames(dds_sva)) in command line}

\item{nickname}{Character string - what do you want these results to be called}

\item{currDate}{Character string - code representing date on which you did this analysis}

\item{expName}{Character string - code representing this whole set of analyses}

\item{genesOfInterest}{Dataframe column listing genes you might want to count in plot, e.g. sex marker genes}
}
\description{
Returns DESeq results object.
}
\details{
This one takes in a coefficient from matrix(resultsNames(dds_sva)). There is also deseq_contrasts if you want to compare specific known groups.
}
\examples{
results_SexEffect_mainEffect <- deseq_contrasts(dds_sva, 0.05, 'SexF.TimepointSecond.Trimester', 'results_SexEffect_mainEffect', '20180814', 'exploreNewData', amySexGenes[,2])
}
\keyword{DESeq}
