% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_contrasts.R
\name{deseq_contrasts}
\alias{deseq_contrasts}
\title{Perform DESeq using contrasts}
\usage{
deseq_contrasts(dds, myAlpha, myContrast, nickname, currDate, expName,
  genesOfInterest)
}
\arguments{
\item{dds}{DESeq object}

\item{myAlpha}{number - threshold for calling genes significant}

\item{myContrast}{c('Factor To Study in Contrast', 'Level1', 'Level2')}

\item{nickname}{Character string - what do you want these results to be called}

\item{currDate}{Character string - code representing date on which you did this analysis}

\item{expName}{Character string - code representing this whole set of analyses}

\item{genesOfInterest}{Dataframe column listing genes you might want to count in plot, e.g. sex marker genes}
}
\description{
You could just call DESeq() in your code... or you could call this nifty function, which 
also saves the output to a table AND an R variable for later use
}
\details{
Returns DESeq results object.
}
\examples{
results_SexEffect_mainEffect <- deseq_contrasts(dds_sva, 0.05, c("Sex", "F", "M"), 'results_SexEffect_mainEffect', '20180814', 'exploreNewData', amySexGenes[,2])
}
\keyword{DESeq}
