% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPCA.R
\name{distPCA}
\alias{distPCA}
\title{PCA Plot}
\usage{
distPCA(cds.pca, condition, colorCondName, PC_A, PC_B, title)
}
\arguments{
\item{cds.pca}{x field of cds.pca, output of prcomp() function}

\item{condition}{column of sampleTable data frame (required for DESeq import) with the factor you want to use to color points}

\item{colorCondName}{Character string - what do you want the point-coloring factor to be called?}

\item{PC_A}{Character string - which PC to be plotted on X axis, e.g. 'PC1'}

\item{PC_B}{Character string - which PC to be plotted on Y axis, e.g. 'PC2'}

\item{title}{Character string - title of plot}
}
\description{
Plots a principal components analysis you've already performed on your data.
}
\keyword{Analysis}
\keyword{DESeq}
\keyword{Data}
\keyword{Exploratory}
