% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSexGenes.R
\name{heatmapSexGenes}
\alias{heatmapSexGenes}
\title{Heatmap of Sex Marker Genes}
\usage{
heatmapSexGenes(myData_rlg, hugoOrEnsembl, sampleTable, myTitle)
}
\arguments{
\item{myData_rlg}{An rlog-normalized DESeqDataset object produced using DESeq's DESeqDataSetFromHTSeqCount() and rlog() functions}

\item{hugoOrEnsembl}{Either a 1 (for HUGO annotations) or 2 (for Ensembl annotations)}

\item{sampleTable}{sampleTable used to import data into DESeq's DESeqDataSetFromHTSeqCount() function}

\item{myTitle}{The title that will display over the plot}
}
\description{
This function allows you to make a heatmap of sex marker gene expression for the samples in your study to make sure you categorized their sex correctly.
Pro tip: male samples are more likely to have high X marker expression and high Y marker expression because of maternal contamination.
}
\examples{
heatmapSexGenes(amySexGenes_Counts, sampleTable, 'Expression of Sex Genes in Samples')

}
\keyword{DESeq2}
\keyword{Heatmap}
