% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVolcanoPlots.R
\name{makeVolcanoPlot}
\alias{makeVolcanoPlot}
\title{Make a volcano plot}
\usage{
makeVolcanoPlot(volcanoPlotData, genesToHighlight, highlightGeneName,
  myTitle, myX, myY, currDate, nickname)
}
\arguments{
\item{volcanoPlotData}{a data frame (preferably of the DESeq results output file). Contains columns named log2FoldChange, padj; and has row names that are genes}

\item{genesToHighlight}{Genes that you'd like to color blue on this plot}

\item{highlightGeneName}{What do you want to call these other points you're coloring blue?}

\item{myTitle}{The title that will display over the plot}

\item{myX}{X axis label}

\item{myY}{Y axis label}
}
\description{
This function allows you to make a volcano plot from your DESeq results.
It plots -log10 of your padjusted values against the log2 fold changes provided by DESeq.
Note that this is not equipped to plot shrunken Log2FCs, just the default DESeq ones - 
but if you label the shrunken log2FCs as "log2FoldChange", the function will plot those instead.
DE genes are colored red, and special other genes of interest colored in blue.
It draws lines at pvalue = 0.05, log2FC = 1.32 (~2.5fold change) and 0.585 (~1.5fold change)
}
\examples{
makeVolcanoPlot(data.frame(results_sex_timepoint), amySexGenes[,2], 'Select Sex Genes', 'DE Genes based on interaction of Timepoint, Sex', 'Log2(Fold Change) relative to E12.5/M', '-log10(Adjusted p-value)')

}
\keyword{DESeq2}
\keyword{RNA-Seq}
\keyword{visualization}
