% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_pickPCs.R
\name{plotPCA_pickPCs}
\alias{plotPCA_pickPCs}
\title{Plot principal components analysis: DESeq2 method, but more than the first two PCs}
\usage{
plotPCA_pickPCs(object, intgroup = "condition", PC_A, PC_B, num_A, num_B,
  ntop = 500, returnData = FALSE)
}
\arguments{
\item{object}{DESeq object, e.g. DESeqTransform}

\item{intgroup}{Character string describing how you want to color the points}

\item{PC_A}{Character string with name of first PC to plot, e.g. 'PC2'}

\item{PC_B}{Character string with name of second PC to plot, e.g. 'PC3'}

\item{num_A}{Integer corresponding to number of PC_A to plot}

\item{num_B}{Integer corresponding to number of PC_B to plot}

\item{ntop}{Number of genes to use when calculating PCs}

\item{returnData}{Return PC data for plotting, or just use this plot?}
}
\description{
This function allows you to plot a Principal Components Analysis with more than the first two PCs.
It uses the same method implemented by DESeq (and is, in fact, an edited version of the plotPC function).
Code courtesy Santosh Anand: https://www.biostars.org/p/243695/
}
\details{
At some point, this code will be rewritten so you don't have to define your PCs of interest twice.
}
\examples{
plotPCA_pickPCs(dds_kal_vst_bc, 'PCR_integration', 'PC2', 'PC3', 2, 3, returnData=TRUE)

}
\keyword{DESeq2,}
\keyword{PCA}
\keyword{RNA-Seq}
\keyword{visualization}
