% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampSimilarityHeatmap.R
\name{sampSimilarityHeatmap}
\alias{sampSimilarityHeatmap}
\title{Examine sample similarity with heatmap}
\usage{
sampSimilarityHeatmap(rawCounts, sampleTable, factorsToPlot,
  sampleTableNameCol)
}
\arguments{
\item{rawCounts}{A dataframe representing counts in your dataset. Could be normalized (rlog(DESeqobject)) or not (counts(DESeqObject))}

\item{sampleTable}{Sample table used to import data in DESeq's DESeqDataSetFromHTSeqCount function}

\item{factorsToPlot}{character list of sampleTable column names you want to plot alongside heatmap}

\item{sampleTableNameCol}{the column of sampleTable with the sample names you want plotted. MUST MATCH COLNAMES OF rawCounts}
}
\description{
This function helps you identify which samples are correlated with one another.
Alongside a heatmap representation of inter-sample correlation, it plots colorful bars representing factors in your sampleTable.
This helps you identify by eye if your samples are clustering according to a certain factor, e.g. sample Sex.
}
\examples{
sampSimilarityHeatmap(data.frame(counts(myData)), sampleTable, c('Sex', 'Timepoint', 'Condition'), sampleTable$FileName)
}
\keyword{Analysis}
\keyword{DESeq2}
\keyword{Data}
\keyword{Exploratory}
