% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcSortBarPlot.R
\name{theoBarPlot}
\alias{theoBarPlot}
\title{The Theo Palmer Bar Plot}
\usage{
theoBarPlot(myResults, highlightGenes, highlightGeneName, fcThreshold_plot,
  fcThreshold, myTitle, myWidth, myHeight)
}
\arguments{
\item{myResults}{Results output from DESeq (not a data frame, but a DESeq object)}

\item{highlightGenes}{column of data frame with genes you want to highlight}

\item{fcThreshold}{Where do you want to draw horizontal line on plot? 0.58 ~= 1.5fold, 1~=2fold, 1.322~=2.5fold}

\item{myTitle}{Character string representing title of plot}

\item{highlightGeneNames}{Character string representing what you want highlighted genes to be called in legend}
}
\description{
Not only does this plot all of your differentially expressed genes as a bar plot with the Y axis representing 
log2fold change, but it also orders all of the bar from positive to negative, highlights the bar corresponding to
a list of interest, and draws a line at a threshold of your choosing.
}
\details{
NOTE: if you want to do this with SHRUNKEN LFCs, NOT THE DESEQ DEFAULT LOG2FCs, use the function theoBarPlot_lfcShrink
}
\examples{
theoBarPlot(results_sex, amySexGenes[,1], 0.58, 0, 'Log2-Fold Change of Genes that are DE by Sex', 15, 6)
}
\keyword{DESeq}
\keyword{Results}
\keyword{Visualization}
