% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoBarPlot_lfcShrink.R
\name{theoBarPlot_lfcShrink}
\alias{theoBarPlot_lfcShrink}
\title{The Theo Palmer Bar Plot ~for shrunken lfcs~}
\usage{
theoBarPlot_lfcShrink(myResults, highlightGenes, highlightGeneName,
  fcThreshold, myTitle)
}
\arguments{
\item{myResults}{Results output from DESeq (can be a data frame too I suppose, e.g. output of add_lfcShrink)}

\item{highlightGenes}{column of data frame with genes you want to highlight}

\item{fcThreshold}{Where do you want to draw horizontal line on plot?}

\item{myTitle}{Character string representing title of plot}

\item{highlightGeneNames}{Character string representing what you want highlighted genes to be called in legend}
}
\description{
Like theoBarPlot, but for shrunken LFCs.
}
\details{
Doesn't plot lfc standard error (yet).

Not only does this plot all of your differentially expressed genes as a bar plot with the Y axis representing 
SHRUNKEN log2fold change, but it also orders all of the bar from positive to negative, highlights the bar corresponding to
a list of interest, and draws a line at a threshold of your choosing.
}
\examples{
theoBarPlot(results_sex, amySexGenes[,1], 0.58, 'Log2-Fold Change of Genes that are DE by Sex')
}
\keyword{DESeq}
\keyword{Results}
\keyword{Visualization}
